forge/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
bpython==0.24
certifi==2024.2.2
charset-normalizer==3.3.2
curtsies==0.4.2
cwcwidth==0.1.9
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
multidict==6.0.5
pygments==2.17.2
pyxdg==0.28
requests==2.31.0
six==1.16.0
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
